simplevideomark: fix display of videomark partially or fully outside video
In case of the videomark being partially or fully outside, an error was being
thrown saying the mark width is more than video width. And when the width,
offset properties are set to maximum it resulted in crash. Instead of throwing
an error, add logic to detect the mark in case of partial visibility or don't
show the mark when it is outside.

https://bugzilla.gnome.org/show_bug.cgi?id=743908