configure: improve check for wayland-scanner binary
The plugin doesn't need the wayland-scanner package to be built
or run, it only needs the wayland-scanner program during compile time.

When cross-compiling, build systems might not have the wayland-scanner
package for the target system as it is a developer's tool, while it should
still be possible to use wayland-scanner from the host system.

This patch fixes it by not requiring the wayland-scanner package but
just the binary itself.

Note that the check is done outside of the PKG_CHECK_MODULES
as it doesn't work inside of it.

https://bugzilla.gnome.org/show_bug.cgi?id=752688